server/requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
colorama==0.4.6
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
sqlparse==0.5.0
tzdata==2024.1
